requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.51.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.31.1
